PM / OPP: Introduce OPP domains for regulator and clock control
As of commit 6a0712f6f199 ("PM / OPP: Add dev_pm_opp_set_rate()") the
OPP core has taken over responsibility for scaling both the clock and
regulator directly for a device when an OPP is changed. Introduce OPP
domains which encapsulate any device clocks and regulators that can be
controlled as needed by the dev_pm_opp_domain_* API exclusively by the
OPP core rather than controlling the clocks and regulators directly.

The current implementation of OPP domains simply takes what was done
before in the OPP core with one clock and one regulator and does a
simple voltage and frequency transition. By moving this into an OPP
domain we create an abstraction that lays the groundwork to provide
platform OPP domain drivers where there are sequencing requirements
for multiple regulators and other non-standard situations. This patch
adds the OPP domain code but does not yet plug it in to the OPP core.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>