CIP: Add a number to the version suffix and move it out of Makefile
We can potentially have more than one CIP release per kernel.org
stable release, so make the version suffix -cip1.

Defining the version suffix in the Makefile resulted in a conflict
when I merged the linux-4.4.y branch, and that's going to happen in
every future merge too.  Instead of defining it there, put it in
localversion-cip.

Signed-off-by: Ben Hutchings <ben.hutchings@codethink.co.uk>