ARM: dts: socfpga: Rename socfpga_cyclone5_de0_{sockit, nano_soc}
Rename DT source for DE0 Nano SoC . The board name is really DE0-Nano-SoC
or Atlas SoC, and it is not to be confused with SoCkit board, which is a
different one. Rename the DT source file to match the board name and to
avoid this possible mixup with another different board.

Upstream commit 202eb5481421040e115526b75317b5ca72584806

This commit is needed because it aligns the device tree name of U-Boot
and kernel which helps distro boot procedures (fdtfile as set by U-Boot
will match the kernel-provided dtb).

Signed-off-by: Marek Vasut <marex@denx.de>
Cc: Dinh Nguyen <dinguyen@kernel.org>
Cc: Jan Kiszka <jan.kiszka@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>
Signed-off-by: Nobuhiro Iwamatsu <nobuhiro.iwamatsu@cybertrust.co.jp>